[#83] Refactor: html, js 분리
html파일을 하나로 통합관리할 경우 todo list 화면에서 페이지 새로고침시 login화면으로 돌아가는 문제 발생

index.html -> login.html, todo.html로 분리
main.js -> login.js, todo.js로 분리
