+++ FE/config/webpack.config.dev.js
@@ -2,12 +2,16 @@ const path = require("path");
 const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 const HtmlWebpackPlugin = require("html-webpack-plugin");
 
-const ENTRY_FILE = path.resolve(__dirname, "../src", "js", "main.js");
+const ENTRY_FILE_LOGIN = path.resolve(__dirname, "../src", "js", "login.js");
+const ENTRY_FILE_TODO = path.resolve(__dirname, "../src", "js", "todo.js");
 const OUTPUT_DIR = path.resolve(__dirname, "../src", "static");
 
 const config = {
 	mode: "development",
-	entry: ENTRY_FILE,
+	entry: {
+		login: ENTRY_FILE_LOGIN,
+		todo: ENTRY_FILE_TODO,
+	},
 	resolve: {
 		alias: {
 			Scss: path.resolve(__dirname, "../src/scss/"),
@@ -40,8 +44,16 @@ const config = {
 			filename: "style.css",
 		}),
 		new HtmlWebpackPlugin({
-			template: path.join(__dirname, "../src/index.html"),
+			filename: "login.html",
+			template: path.join(__dirname, "../src/login.html"),
 			hash: true,
+			chunks: ["login"],
+		}),
+		new HtmlWebpackPlugin({
+			filename: "todo.html",
+			template: path.join(__dirname, "../src/todo.html"),
+			hash: true,
+			excludeChunks: ["login"],
 		}),
 	],
 	output: {
