+++ FE/src/login.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+	<head>
+		<meta charset="UTF-8" />
+		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+		<title>Todo List</title>
+	</head>
+	<body>
+		<div class="login-container">
+			<h1>Todo List</h1>
+			<button class="login">LOG IN</button>
+		</div>
+	</body>
+</html>
