+++ FE/src/todo.html
@@ -7,9 +7,6 @@
 		<script src="https://kit.fontawesome.com/67afa37cd7.js" crossorigin="anonymous"></script>
 	</head>
 	<body>
-		<div class="login-container">
-			<h1>Todo List</h1>
-			<button class="login">LOG IN</button>
-		</div>
+		<div id="Todo"></div>
 	</body>
 </html>
